tiltfile: add the set() built-in. Fixes https://github.com/tilt-dev/tilt/issues/3773
